pyproject.toml
[project]
name = "remove-background"
version = "0.1.0"
description = "Local RMBG-2.0 background removal CLI and HTTP API."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "kornia>=0.7.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.9",
    "timm>=1.0.0",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "transformers>=4.44.0",
    "typer>=0.12.0",
    "uvicorn[standard]>=0.30.0",
]

[project.scripts]
rmbg = "remove_background.cli:app"

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
